Remove 8 locations from "suppressed" to allow holds
Based on conversation with Research staff, this removes 8 recap customer
codes from the array of "suppressed" locations so that the app creates a
hold on those items. A recap customer code's presense in the
SUPPRESSION_CODES array causes this app to *skip* placing a hold on any
item destined for those locations.

https://jira.nypl.org/browse/SCC-2642