Fix bug handling hold failures at staff locations
Fixes bug where hold failures at staff-only locations result in an error
response, causing the request at SCSB to be rejected. Hold failures at
staff-only locations are almost certainly the result of the item being
suppressed. (And if there's another reason for the failure, we should
allow the item to be retrieved anyway because it's bound for a staff
only location, where staff will know how to handle it.)

https://jira.nypl.org/browse/SCC-2748